Bug 1815757 - limit size of InterfaceOffset to reduce padding r=xpcom-reviewers,nika
Differential Revision: https://phabricator.services.mozilla.com/D169230

UltraBlame original commit: 5a2d981bf8ce41a7bb70dbf3135e5967fd7a15e0

+++ xpcom/components/StaticComponents.h
@@ -84,7 +84,7 @@ struct StringOffset final {
 
 
 struct InterfaceOffset final {
-  uint32_t mOffset;
+  uint16_t mOffset;
 };
 
 
@@ -223,6 +223,7 @@ struct JSServiceEntry final {
   ModuleID mModuleID;
 
   InterfaceOffset mInterfaceOffset;
+
   uint8_t mInterfaceCount;
 
   nsCString Name() const;
